requirements.txt
requests
beautifulsoup4
lxml
extruct
w3lib
tenacity
tqdm
pillow
pandas
numpy
